refactor: move SdramCtrlTester
